ingestion/requirements.txt
requests==2.28.1
kafka-python==2.0.2
python-dateutil==2.8.2
retry==0.9.2